requirements.txt
alembic==1.4.0
attrs==19.3.0
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.2
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.8
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
dominate==2.4.0
elasticsearch==7.5.1
Flask==1.1.1
Flask-BasicAuth==0.2.0
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.16.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
Mako==1.1.1
MarkupSafe==1.1.1
paramiko==2.7.1
pathspec==0.5.9
pycparser==2.19
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
python-dotenv==0.11.0
python-editor==1.0.4
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
SQLAlchemy==1.3.13
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
visitor==0.1.3
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==1.0.0
WTForms==2.2.1
